Use 3' read file instead of taginfo file
Tailseeker makes some questionable poly-A length calls; to correct these,
we interrogate the processed 3' file. This file also contains the
remaining tag information in its read IDs. Furthermore, there is some
unexplained mismatch between the rows in the taginfo file and the fastq
file. We therefore skip the taginfo file completely and go directly to
the source.