Verify 3p reads mapping to the viral RNA
Putative matches based on mapping 5p reads have a tendency to be
anywhere but near the 3p end of the viral RNA. This is corrected here by
manually aligning the 3p end of the viral RNA reference against every
candidate 3p tail read and only preserving close matches.

In the same vein we also manually determine poly(A) tailing and tail
modifications here, since the ones created by Tailseeker appear to be
entirely random when compared to the reference-based approach taken
here.